Move B1520A module to its own module KeysightB1520A.py
